start of `the big protocol breakup`
